pyproject.toml
[project]
name            = "dspy-lora-agent"
version         = "0.1.0"
description     = "Backend for cardiac fibrosis compound selection system."
readme          = "README.md"
requires-python = ">=3.12"
dependencies = [
    "modal",
    "dspy",
    "tavily-python",
    "python-dotenv",
    "openai==1.99.5", # pinned becaused of bug https://github.com/stanfordnlp/dspy/issues/8677
    "datasets>=2.14.6", # copied version used in dspy
    "huggingface"
]

[build-system]
requires      = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.ruff]
exclude = []

[tool.ruff.per-file-ignores]
"agentic_system/agents/*.py" = ["E402"]